Moved sub-classes of "DataGridViewRowCollection", "ListBox.IntegerCollection" and "ListBox.SelectedIndexCollection" classes to their own files
