+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewRowCollection.cs
@@ -20,7 +20,7 @@ namespace System.Windows.Forms
     [ListBindable(false)]
     [DesignerSerializer("System.Windows.Forms.Design.DataGridViewRowCollectionCodeDomSerializer, " + AssemblyRef.SystemDesign,
                         "System.ComponentModel.Design.Serialization.CodeDomSerializer, " + AssemblyRef.SystemDesign)]
-    public class DataGridViewRowCollection : ICollection, IList
+    public partial class DataGridViewRowCollection : ICollection, IList
     {
 #if DEBUG
         // set to false when the cached row heights are dirty and should not be accessed.
@@ -2469,309 +2469,6 @@ namespace System.Windows.Forms
             cachedRowCountsAccessAllowed = true;
             cachedRowHeightsAccessAllowed = true;
 #endif
-        }
-
-        private class RowArrayList : ArrayList
-        {
-            private readonly DataGridViewRowCollection owner;
-            private RowComparer rowComparer;
-
-            public RowArrayList(DataGridViewRowCollection owner)
-            {
-                this.owner = owner;
-            }
-
-            public void CustomSort(RowComparer rowComparer)
-            {
-                Debug.Assert(rowComparer is not null);
-                Debug.Assert(Count > 0);
-
-                this.rowComparer = rowComparer;
-                CustomQuickSort(0, Count - 1);
-            }
-
-            private void CustomQuickSort(int left, int right)
-            {
-                // Custom recursive QuickSort needed because of the notion of shared rows.
-                // The indexes of the compared rows are required to do the comparisons.
-                // For a study comparing the iterative and recursive versions of the QuickSort
-                // see https://web.archive.org/web/20051125015050/http://www.mathcs.carleton.edu/courses/course_resources/cs227_w96/wightmaj/data.html
-                // Is the recursive version going to cause trouble with large dataGridViews?
-                do
-                {
-                    if (right - left < 2) // sort subarray of two elements
-                    {
-                        if (right - left > 0 && rowComparer.CompareObjects(rowComparer.GetComparedObject(left), rowComparer.GetComparedObject(right), left, right) > 0)
-                        {
-                            owner.SwapSortedRows(left, right);
-                        }
-                        return;
-                    }
-
-                    int k = (left + right) >> 1;
-                    object x = Pivot(left, k, right);
-                    int i = left + 1;
-                    int j = right - 1;
-                    do
-                    {
-                        while (k != i && rowComparer.CompareObjects(rowComparer.GetComparedObject(i), x, i, k) < 0)
-                        {
-                            i++;
-                        }
-                        while (k != j && rowComparer.CompareObjects(x, rowComparer.GetComparedObject(j), k, j) < 0)
-                        {
-                            j--;
-                        }
-                        Debug.Assert(i >= left && j <= right, "(i>=left && j<=right)  Sort failed - Is your IComparer bogus?");
-                        if (i > j)
-                        {
-                            break;
-                        }
-                        if (i < j)
-                        {
-                            owner.SwapSortedRows(i, j);
-                            if (i == k)
-                            {
-                                k = j;
-                            }
-                            else if (j == k)
-                            {
-                                k = i;
-                            }
-                        }
-                        i++;
-                        j--;
-                    }
-                    while (i <= j);
-
-                    if (j - left <= right - i)
-                    {
-                        if (left < j)
-                        {
-                            CustomQuickSort(left, j);
-                        }
-                        left = i;
-                    }
-                    else
-                    {
-                        if (i < right)
-                        {
-                            CustomQuickSort(i, right);
-                        }
-                        right = j;
-                    }
-                }
-                while (left < right);
-            }
-
-            private object Pivot(int left, int center, int right)
-            {
-                // find median-of-3 (left, center and right) and sort these 3 elements
-                if (rowComparer.CompareObjects(rowComparer.GetComparedObject(left), rowComparer.GetComparedObject(center), left, center) > 0)
-                {
-                    owner.SwapSortedRows(left, center);
-                }
-                if (rowComparer.CompareObjects(rowComparer.GetComparedObject(left), rowComparer.GetComparedObject(right), left, right) > 0)
-                {
-                    owner.SwapSortedRows(left, right);
-                }
-                if (rowComparer.CompareObjects(rowComparer.GetComparedObject(center), rowComparer.GetComparedObject(right), center, right) > 0)
-                {
-                    owner.SwapSortedRows(center, right);
-                }
-                return rowComparer.GetComparedObject(center);
-            }
-        }
-
-        private class RowComparer
-        {
-            private readonly DataGridView dataGridView;
-            private readonly DataGridViewRowCollection dataGridViewRows;
-            private readonly DataGridViewColumn dataGridViewSortedColumn;
-            private readonly int sortedColumnIndex;
-            private readonly IComparer customComparer;
-            private readonly bool ascending;
-            private static readonly ComparedObjectMax max = new ComparedObjectMax();
-
-            public RowComparer(DataGridViewRowCollection dataGridViewRows, IComparer customComparer, bool ascending)
-            {
-                dataGridView = dataGridViewRows.DataGridView;
-                this.dataGridViewRows = dataGridViewRows;
-                dataGridViewSortedColumn = dataGridView.SortedColumn;
-                if (dataGridViewSortedColumn is null)
-                {
-                    Debug.Assert(customComparer is not null);
-                    sortedColumnIndex = -1;
-                }
-                else
-                {
-                    sortedColumnIndex = dataGridViewSortedColumn.Index;
-                }
-                this.customComparer = customComparer;
-                this.ascending = ascending;
-            }
-
-            internal object GetComparedObject(int rowIndex)
-            {
-                if (dataGridView.NewRowIndex != -1)
-                {
-                    Debug.Assert(dataGridView.AllowUserToAddRowsInternal);
-                    if (rowIndex == dataGridView.NewRowIndex)
-                    {
-                        return max;
-                    }
-                }
-                if (customComparer is null)
-                {
-                    DataGridViewRow dataGridViewRow = dataGridViewRows.SharedRow(rowIndex);
-                    Debug.Assert(dataGridViewRow is not null);
-                    Debug.Assert(sortedColumnIndex >= 0);
-                    return dataGridViewRow.Cells[sortedColumnIndex].GetValueInternal(rowIndex);
-                }
-                else
-                {
-                    return dataGridViewRows[rowIndex]; // Unsharing compared rows!
-                }
-            }
-
-            internal int CompareObjects(object value1, object value2, int rowIndex1, int rowIndex2)
-            {
-                if (value1 is ComparedObjectMax)
-                {
-                    return 1;
-                }
-                else if (value2 is ComparedObjectMax)
-                {
-                    return -1;
-                }
-                int result = 0;
-                if (customComparer is null)
-                {
-                    if (!dataGridView.OnSortCompare(dataGridViewSortedColumn, value1, value2, rowIndex1, rowIndex2, out result))
-                    {
-                        if (!(value1 is IComparable) && !(value2 is IComparable))
-                        {
-                            if (value1 is null)
-                            {
-                                if (value2 is null)
-                                {
-                                    result = 0;
-                                }
-                                else
-                                {
-                                    result = 1;
-                                }
-                            }
-                            else if (value2 is null)
-                            {
-                                result = -1;
-                            }
-                            else
-                            {
-                                result = Comparer.Default.Compare(value1.ToString(), value2.ToString());
-                            }
-                        }
-                        else
-                        {
-                            result = Comparer.Default.Compare(value1, value2);
-                        }
-                        if (result == 0)
-                        {
-                            if (ascending)
-                            {
-                                result = rowIndex1 - rowIndex2;
-                            }
-                            else
-                            {
-                                result = rowIndex2 - rowIndex1;
-                            }
-                        }
-                    }
-                }
-                else
-                {
-                    Debug.Assert(value1 is DataGridViewRow);
-                    Debug.Assert(value2 is DataGridViewRow);
-                    Debug.Assert(value1 is not null);
-                    Debug.Assert(value2 is not null);
-                    //
-
-                    result = customComparer.Compare(value1, value2);
-                }
-
-                if (ascending)
-                {
-                    return result;
-                }
-                else
-                {
-                    return -result;
-                }
-            }
-
-            class ComparedObjectMax
-            {
-                public ComparedObjectMax() { }
-            }
-        }
-
-        private class UnsharingRowEnumerator : IEnumerator
-        {
-            private readonly DataGridViewRowCollection owner;
-            private int current;
-
-            /// <summary>
-            ///  Creates a new enumerator that will enumerate over the rows and unshare the accessed rows if needed.
-            /// </summary>
-            public UnsharingRowEnumerator(DataGridViewRowCollection owner)
-            {
-                this.owner = owner;
-                current = -1;
-            }
-
-            /// <summary>
-            ///  Moves to the next element, or returns false if at the end.
-            /// </summary>
-            bool IEnumerator.MoveNext()
-            {
-                if (current < owner.Count - 1)
-                {
-                    current++;
-                    return true;
-                }
-                else
-                {
-                    current = owner.Count;
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Resets the enumeration back to the beginning.
-            /// </summary>
-            void IEnumerator.Reset()
-            {
-                current = -1;
-            }
-
-            /// <summary>
-            ///  Retrieves the current value in the enumerator.
-            /// </summary>
-            object IEnumerator.Current
-            {
-                get
-                {
-                    if (current == -1)
-                    {
-                        throw new InvalidOperationException(SR.DataGridViewRowCollection_EnumNotStarted);
-                    }
-                    if (current == owner.Count)
-                    {
-                        throw new InvalidOperationException(SR.DataGridViewRowCollection_EnumFinished);
-                    }
-                    return owner[current];
-                }
-            }
         }
     }
 }
